2024 - Day 20 - Optimize part one.
Previously it took 42 mins to brute force every map.  Instead, since we know the original map has a single path, we know the distance from any point on that path to the end, so we can just measure the difference from the start of the cheat to the end of it.